Removed the user stats page and ported the historical pages/words by year to the new profile component.
