--- UI/Web/src/app/statistics/_components/user-stats/user-stats.component.html
@@ -1,25 +1,0 @@
-<ng-container *transloco="let t; prefix:'user-stats'">
-  @if (userId) {
-    <p class="alert alert-warning" role="alert">
-      This page is being removed, all stats will be on <a href="/profile/{{userId}}">profile page</a> going forward
-    </p>
-    <div class="vstack gap-3">
-      <div class="row g-0 d-flex justify-content-around">
-        @if (userStats$ | async; as userStats) {
-          <app-user-stats-info-cards [totalPagesRead]="userStats.totalPagesRead" [totalWordsRead]="userStats.totalWordsRead" [timeSpentReading]="userStats.timeSpentReading"
-                                     [chaptersRead]="userStats.chaptersRead" [lastActive]="userStats.lastActiveUtc" [avgHoursPerWeekSpentReading]="userStats.avgHoursPerWeekSpentReading" />
-        }
-      </div>
-<!--      <div class="row g-0 fixed-row">-->
-<!--        <app-reading-activity [userId]="userId" [isAdmin]="(accountService.isAdmin$ | async) || false" [individualUserMode]="true" />-->
-<!--      </div>-->
-<!--      <div class="row g-0 fixed-row">-->
-<!--        <app-day-breakdown [userId]="userId" />-->
-<!--      </div>-->
-<!--      <div class="row g-0 fixed-row">-->
-<!--        <app-stat-list [data$]="percentageRead$" [label]="t('read-percentage')" [title]="t('library-read-progress-title')" />-->
-<!--      </div>-->
-    </div>
-  }
-
-</ng-container>
