--- UI/Web/src/app/statistics/_components/user-stats/user-stats.component.scss
@@ -1,3 +1,0 @@
-.fixed-row {
-  height: auto; // 242px;
-}
